+++ lib/pdataframe.dart
@@ -1,3 +1,4 @@
 export 'src/dataframe.dart';
 export 'src/dfunctions.dart';
 export 'src/timestamp.dart';
+export 'src/math_utils.dart';